HEEDLS-585 Pull out common DelegatesVMFilterOptions helper functions
