+++ DigitalLearningSolutions.Web/ViewModels/TrackingSystem/Delegates/EmailDelegates/EmailDelegatesViewModelFilterOptions.cs
@@ -3,58 +3,11 @@
     using System.Collections.Generic;
     using System.Linq;
     using DigitalLearningSolutions.Data.Models.CustomPrompts;
-    using DigitalLearningSolutions.Data.Models.User;
-    using DigitalLearningSolutions.Web.Helpers;
-    using DigitalLearningSolutions.Web.Models.Enums;
     using DigitalLearningSolutions.Web.ViewModels.Common.SearchablePage;
+    using DigitalLearningSolutions.Web.ViewModels.TrackingSystem.Delegates.Shared;
 
     public static class EmailDelegatesViewModelFilterOptions
     {
-        public static IEnumerable<FilterOptionViewModel> GetJobGroupOptions(
-            IEnumerable<(int id, string name)> jobGroups
-        )
-        {
-            return jobGroups.Select(
-                jobGroup => new FilterOptionViewModel(
-                    jobGroup.name,
-                    FilteringHelper.BuildFilterValueString(
-                        nameof(DelegateUserCard.JobGroupId),
-                        nameof(DelegateUserCard.JobGroupId),
-                        jobGroup.id.ToString()
-                    ),
-                    FilterStatus.Default
-                )
-            );
-        }
-
-        public static IEnumerable<FilterOptionViewModel> GetCustomPromptOptions(
-            CustomPrompt customPrompt
-        )
-        {
-            string filterValueName =
-                CentreCustomPromptHelper.GetDelegateCustomPromptAnswerName(customPrompt.CustomPromptNumber);
-
-            var options = customPrompt.Options.Select(
-                option => new FilterOptionViewModel(
-                    option,
-                    FilteringHelper.BuildFilterValueString(filterValueName, filterValueName, option),
-                    FilterStatus.Default
-                )
-            ).ToList();
-            options.Add(
-                new FilterOptionViewModel(
-                    "No option selected",
-                    FilteringHelper.BuildFilterValueString(
-                        filterValueName,
-                        filterValueName,
-                        FilteringHelper.EmptyValue.ToString()
-                    ),
-                    FilterStatus.Default
-                )
-            );
-            return options;
-        }
-
         public static List<FilterViewModel> GetEmailDelegatesFilterViewModels(
             IEnumerable<(int id, string name)> jobGroups,
             IEnumerable<CustomPrompt> promptsWithOptions
@@ -65,7 +18,7 @@
                 new FilterViewModel(
                     "JobGroupId",
                     "Job Group",
-                    GetJobGroupOptions(jobGroups)
+                    DelegatesViewModelFilterOptions.GetJobGroupOptions(jobGroups)
                 )
             };
             filters.AddRange(
@@ -73,7 +26,7 @@
                     customPrompt => new FilterViewModel(
                         $"CustomPrompt{customPrompt.CustomPromptNumber}",
                         customPrompt.CustomPromptText,
-                        GetCustomPromptOptions(customPrompt)
+                        DelegatesViewModelFilterOptions.GetCustomPromptOptions(customPrompt)
                     )
                 )
             );
